+++ fabric/src/main/java/com/lx862/jcm/mod/extra/SignExtraAccessor.java
@@ -1,0 +1,5 @@
+package com.lx862.jcm.mod.extra;
+
+public interface SignExtraAccessor {
+    int jsblock$getSignTextColor();
+}